requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.13.0
google-cloud-secret-manager==2.16.4
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
